requirements.txt
opencv-python
ultralytics
easyocr
pandas
python-dotenv
requests
